middleware/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
websockets==12.0
python-dotenv==1.0.1
httpx>=0.28.1
twilio>=9.0.0
google-genai
google-auth
